app/requirements.txt
astroquery==0.4.3
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
astropy==4.3.1
pytest==6.2.5
python-dotenv==0.19.2